+++ src/Bootstrap.cpp
@@ -53,6 +53,7 @@ Rcpp::List LSVCMM_Bootstrap(
     double backtracking_fraction,
     bool two_step_estimation,
     double stepsize_factor,
+    bool update_stepsize_every_round,
     unsigned int n_samples,
     bool resample_within_subject
 ){
@@ -64,7 +65,9 @@ Rcpp::List LSVCMM_Bootstrap(
     update_method,
     backtracking_fraction,
     two_step_estimation,
-    stepsize_factor
+    10.,
+    stepsize_factor,
+    update_stepsize_every_round
   );
 
   if(control->verbose) Rcpp::Rcout << "[LSVCMM] Initializing data \n";