+++ src/Control.cpp
@@ -17,7 +17,9 @@ public:
   std::string update_method = "PGD";
   double backtracking_fraction = 0.9;
   bool two_step_estimation = true;
+  double lambda_max_factor = 10;
   double stepsize_factor = 0.1;
+  bool update_stepsize_every_round = false;
 
   Control(
     uint max_rounds,
@@ -27,7 +29,9 @@ public:
     std::string update_method,
     double backtracking_fraction,
     bool two_step_estimation,
-    double stepsize_factor
+    double lambda_max_factor,
+    double stepsize_factor,
+    bool update_stepsize_every_round
   ){
     this->max_iter = max_iter;
     this->max_rounds = max_rounds;
@@ -36,7 +40,9 @@ public:
     this->update_method = update_method;
     this->backtracking_fraction = backtracking_fraction;
     this->two_step_estimation = two_step_estimation;
+    this->lambda_max_factor = lambda_max_factor;
     this->stepsize_factor = stepsize_factor;
+    this->update_stepsize_every_round = update_stepsize_every_round;
   }
 
   Control(){}
@@ -49,7 +55,9 @@ public:
     results["control.update_method"] = this->update_method;
     results["control.backtracking_fraction"] = this->backtracking_fraction;
     results["control.two_step_estimation"] = this->two_step_estimation;
+    results["control.lambda_max_factor"] = this->lambda_max_factor;
     results["control.stepsize_factor"] = this->stepsize_factor;
+    results["control.update_stepsize_every_round"] = this->update_stepsize_every_round;
   }
 
 };